+++ editor_api/actions/setup_project.py
@@ -21,7 +21,7 @@ from playhouse.migrate import *
 OVERWRITE_PLANTS = False
 
 class SetupProject(ExecutableApi):
-	def __init__(self, project_db, editor_version, settings_file=None, project_name=None, datasets_db=None, constant_ps=True, is_lte=False):
+	def __init__(self, project_db, editor_version, project_name=None, datasets_db=None, constant_ps=True, is_lte=False, project_description=None):
 		self.__abort = False
 
 		base_path = os.path.dirname(project_db)
@@ -36,7 +36,8 @@ class SetupProject(ExecutableApi):
 			try:
 				config = Project_config.get()
 				datasets_db = utils.full_path(project_db, config.reference_db)
-				project_name = config.project_name
+				if project_name is None:
+					project_name = config.project_name
 			except Project_config.DoesNotExist:
 				sys.exit('Could not retrieve project configuration data.')
 
@@ -47,24 +48,28 @@ class SetupProject(ExecutableApi):
 			sys.exit(ver_check)
 
 		# Backup original db before beginning
-		try:
-			self.emit_progress(2, 'Backing up GIS database...')
-			filename, file_extension = os.path.splitext(rel_project_db)
-			bak_filename = filename + '_bak_' + time.strftime('%Y%m%d-%H%M%S') + file_extension
-			bak_dir = os.path.join(base_path, 'DatabaseBackups')
-			if not os.path.exists(bak_dir):
-				os.makedirs(bak_dir)
-			backup_db_file = os.path.join(bak_dir, bak_filename)
-			copyfile(project_db, backup_db_file)
-		except IOError as err:
-			sys.exit(err)
+		do_gis = False
+		if os.path.exists(project_db):
+			do_gis = True
+			try:
+				self.emit_progress(2, 'Backing up GIS database...')
+				filename, file_extension = os.path.splitext(rel_project_db)
+				bak_filename = filename + '_bak_' + time.strftime('%Y%m%d-%H%M%S') + file_extension
+				bak_dir = os.path.join(base_path, 'DatabaseBackups')
+				if not os.path.exists(bak_dir):
+					os.makedirs(bak_dir)
+				backup_db_file = os.path.join(bak_dir, bak_filename)
+				copyfile(project_db, backup_db_file)
+			except IOError as err:
+				sys.exit(err)
 
 		try:
 			SetupProjectDatabase.init(project_db, datasets_db)
 			self.emit_progress(10, 'Creating database tables...')
 			SetupProjectDatabase.create_tables()
 			self.emit_progress(50, 'Copying data from SWAT+ datasets database...')
-			SetupProjectDatabase.initialize_data(project_name, is_lte, overwrite_plants=OVERWRITE_PLANTS)
+			description = project_description if project_description is not None else project_name
+			SetupProjectDatabase.initialize_data(description, is_lte, overwrite_plants=OVERWRITE_PLANTS)
 
 			config = Project_config.get_or_create_default(
 				editor_version=editor_version,
@@ -97,23 +102,9 @@ class SetupProject(ExecutableApi):
 				self.emit_progress(100, "Error occurred.")
 			sys.exit(str(ex))
 
-		api = GisImport(project_db, True, constant_ps, backup_db_file)
-		api.insert_default()
-
-		settings_data = {
-			'swatplus-project': {
-				'version': editor_version,
-				'name': project_name,
-				'databases': {
-					'project': rel_project_db,
-					'datasets': rel_datasets_db
-				},
-				'model': 'SWAT+' if not is_lte else 'SWAT+ lte'
-			}
-		}
-
-		with open(settings_file, 'w') as file:
-			json.dump(settings_data, file, indent='\t')
+		if do_gis:
+			api = GisImport(project_db, True, constant_ps, backup_db_file)
+			api.insert_default()
 
 
 if __name__ == '__main__':
@@ -123,8 +114,7 @@ if __name__ == '__main__':
 	parser.add_argument("--project_name", type=str, help="project name", nargs="?")
 	parser.add_argument("--editor_version", type=str, help="editor version", nargs="?")
 	parser.add_argument("--datasets_db_file", type=str, help="full path of datasets SQLite database file", nargs="?")
-	parser.add_argument("--settings_file", type=str, help="editor version", nargs="?")
-	parser.add_argument("--constant_ps", type=str, help="y/n constant point source values (default y)", nargs="?")
+	parser.add_argument("--constant_ps", type=str, help="y/n constant point source values (default n)", nargs="?")
 	parser.add_argument("--is_lte", type=str, help="y/n use lte version of SWAT+ (default n)", nargs="?")
 
 	args = parser.parse_args()
@@ -132,4 +122,4 @@ if __name__ == '__main__':
 	constant_ps = True if args.constant_ps == "y" else False
 	is_lte = True if args.is_lte == "y" else False
 
-	api = SetupProject(args.project_db_file, args.editor_version, args.settings_file, args.project_name, args.datasets_db_file, constant_ps, is_lte)
+	api = SetupProject(args.project_db_file, args.editor_version, args.project_name, args.datasets_db_file, constant_ps, is_lte)
